Fix placement heap check in `MVKImagePlane::getMTLTexture()`
`MVKDeviceMemory::ensureMTLHeap()` returns false only when a placement heap
is supported but failed allocation. If placement heaps are not supported or
it is not created for some reason, it still returns true.

If we want to check that a placement heap is available for use, then we
need to call `MVKDeviceMemory::getMTLHeap()`. Since `ensureMTLHeap()` is
called at construction (if supported), the pointer will only be non-NULL if
we indeed support and have a placement heap.